refactor: move "nodes" CLI command to commands/ folder
